增加功能

+++ src/main/java/com/mn/modules/api/vo/ArrivedData.java
@@ -14,5 +14,5 @@ public class ArrivedData {
     /**
      * 到访比例
      */
-    Integer arrivedRate;
+    Double arrivedRate;
 }
